Dataset: Fixed bug that caused entries and cvs to get pulled from server when using views Testing: moved df_equals to testing.py
